flask/requirements.txt
vosk
scipy
Flask
python-dotenv
numpy